requirements.txt
transformers
torch
gradio
python-dotenv